Function to automatically buy in bulk (test)
Im trying to code a function to auto select buying in bulk that would be really usefull to players with a lot of prestige. When using auto ascend buying x1 my game needs 5:40 mins to get the first HC but on x100 it only needs 0:45.

Please check that I did not write any nonsense as I not yet fully understand how frozen cookies work.